requirements.txt
Django==2.0.6
django-rest-framework==0.1.0
